% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{adjacency_matrix}
\alias{adjacency_matrix}
\title{Creat adjacency matrix for all the operations in the context}
\usage{
adjacency_matrix(ctx)
}
\arguments{
\item{ctx}{aria_context object}
}
\description{
Creat adjacency matrix for all the operations in the context
}
\examples{
library(aria)
ctx <- get_context()
register_ops(ctx, cars)
register_ops(ctx, data.frame)
x <- cpu_tensor(5, dims = 1)
y <- x ** 3
y <- (x ** 3) / 2
all.ops <- get_all_ops_ptr(ctx)
adj_mat <- adjacency_matrix(ctx)
print(adj_mat)
plot(adj_mat)
}
