% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appIMPACT.R
\name{appIMPACT}
\alias{appIMPACT}
\title{Title}
\usage{
appIMPACT(folder)
}
\arguments{
\item{folder}{Folder path where IMPACT outputs are stored.
Typically of format <path_to_impact_folder>/OutputFiles/Scenarios}
}
\value{
Fires the app for IMPACT runs
}
\description{
Title
}
\examples{
\dontrun{appIMPACT()}
}
\author{
Abhijeet Mishra
}
