% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aria.R
\docType{package}
\name{aria-package}
\alias{aria}
\alias{aria-package}
\title{aria: Extensible Autodifferentiation}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Krzysztof Joachimiak \email{joachimiak.krzysztof@gmail.com} (0000-0003-4780-7947)

}
\keyword{internal}
