% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{ARIA_OPTIONS}
\alias{ARIA_OPTIONS}
\title{List of `aria` options}
\format{An object of class \code{list} of length 2.}
\usage{
ARIA_OPTIONS
}
\description{
List of `aria` options
}
\keyword{datasets}
