% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_syntax.R
\name{check_syntax}
\alias{check_syntax}
\title{Check Stan code for syntax errors}
\usage{
check_syntax(code_path)
}
\arguments{
\item{code_path}{Character string describing the path to the Stan code.}
}
\value{
Returns NULL invisibly if called from the global environment; if called from inside a function, returns a boolean value with TRUE indicating that the check has passed.
}
\description{
Check Stan code for syntax errors
}
\examples{
\dontrun{
check_syntax('my_model.stan')
}
}
\seealso{
Other Model checking & compilation functions: 
\code{\link{check_syntax_and_maybe_compile}()},
\code{\link{compile}()},
\code{\link{disable_rstudio_syntax_compile}()},
\code{\link{enable_rstudio_syntax_compile}()}
}
\concept{Model checking & compilation functions}
