% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_syntax_and_maybe_compile.R
\name{check_syntax_and_maybe_compile}
\alias{check_syntax_and_maybe_compile}
\title{Check a Stan file for syntax errors & possibly compile (see description)}
\usage{
check_syntax_and_maybe_compile(code_path)
}
\arguments{
\item{code_path}{Character string describing the path to the Stan code.}
}
\value{
NULL (invisibly); Side effects: if the model was compiled, an eponymous executable binary is placed in a folder called `aria` (along with some other helper files).
}
\description{
This function calls \code{\link{check_syntax}}; if no syntax errors are detected, the model is then compiled with \code{\link{compile}} \strong{if-and-only-if} the stan file has a first line that is precisely \code{\\compile:1}.
}
\examples{
\dontrun{
check_and_maybe_compile('my_model.stan')
}
}
\seealso{
Other Model checking & compilation functions: 
\code{\link{check_syntax}()},
\code{\link{compile}()},
\code{\link{disable_rstudio_syntax_compile}()},
\code{\link{enable_rstudio_syntax_compile}()}
}
\concept{Model checking & compilation functions}
