% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_all.R
\name{clean_all}
\alias{clean_all}
\title{Delete ALL temporary files}
\usage{
clean_all()
}
\value{
Returns NULL invisibly
}
\description{
Delete ALL temporary files
}
\examples{
\dontrun{
clean_all()
}
}
\seealso{
Other Cleaning functions: 
\code{\link{clean}()}
}
\concept{Cleaning functions}
