% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{coda}
\alias{coda}
\title{Collect output from aria}
\usage{
coda()
}
\value{
a named list with elements `samples`, `stdout`, `stderr`, `info`, & `time`.
}
\description{
Collect output from aria
}
\examples{
\dontrun{
p = aria::coda()
}
}
