% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\title{Collect output from aria}
\usage{
collect()
}
\value{
a named list with elements `samples`, `stdout`, `stderr`, `info`, & `time`.
}
\description{
Collect output from aria
}
\examples{
\dontrun{
p = aria::collect()
}
}
