% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_ptr}
\alias{compare_ptr}
\title{Compare two pointers}
\usage{
compare_ptr(x, y)
}
\arguments{
\item{x}{external pointer}

\item{y}{external pointer}
}
\description{
Compare two pointers
}
\examples{
library(aria)
ctx <- get_context()
register_ops(ctx, cars)
register_ops(ctx, data.frame)
x <- cpu_tensor(5, dims = 1)
all_ops <- get_all_ops_ptr(ctx)
compare_ptr(x@pointer, all_ops[[3]])
}
