% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx-ops.R
\name{connect}
\alias{connect}
\title{Connect tensors and functions to track computational chain.
It's needed to compute gradients}
\usage{
connect(input, output)
}
\arguments{
\item{input}{Input operations (tensors or functions)}

\item{output}{Output operations (tensors or functions)}
}
\description{
TODO: rename input and output
FIXME: it's probably a crucial function!
}
