% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disable_rstudio_syntax_compile.R
\name{disable_rstudio_syntax_compile}
\alias{disable_rstudio_syntax_compile}
\title{Disable enhanced syntax check (inc. compile) in RStuidio}
\usage{
disable_rstudio_syntax_compile()
}
\value{
NULL (invisibly)
}
\description{
This function returns Rstudio (& rstan) to their default behaviour.
}
\examples{
\dontrun{
disable_rstudio_syntax_compile()
}
}
\seealso{
Other Model checking & compilation functions: 
\code{\link{check_syntax_and_maybe_compile}()},
\code{\link{check_syntax}()},
\code{\link{compile}()},
\code{\link{enable_rstudio_syntax_compile}()}
}
\concept{Model checking & compilation functions}
