% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{.abstract_operator}
\alias{.abstract_operator}
\title{Abstract function for tensor-numeric operations
Backward function rather than partial derivative?
Problem: newly created tensor has by default
Maybe we don't need to track some 'abstract tensors'
Memoisation can handle this case instead}
\usage{
.abstract_operator(fun, deriv, x, y)
}
\description{
Abstract function for tensor-numeric operations
Backward function rather than partial derivative?
Problem: newly created tensor has by default
Maybe we don't need to track some 'abstract tensors'
Memoisation can handle this case instead
}
