% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{.addition}
\alias{.addition}
\title{============================================================== #
                             ADDITION                          #
============================================================== #
Problem:
We don't want to recreate functions
Should we use 'temporal' context?}
\usage{
.addition(x, y)
}
\description{
============================================================== #
                             ADDITION                          #
============================================================== #
Problem:
We don't want to recreate functions
Should we use 'temporal' context?
}
