% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{.division}
\alias{.division}
\title{============================================================== #
                             DIVISION                          #
============================================================== #}
\usage{
.division(x, y)
}
\description{
============================================================== #
                             DIVISION                          #
============================================================== #
}
