% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{.modulo}
\alias{.modulo}
\title{============================================================== #
                             MODULO                            #
============================================================== #}
\usage{
.modulo(x, y)
}
\description{
============================================================== #
                             MODULO                            #
============================================================== #
}
