% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{.multiplication}
\alias{.multiplication}
\title{============================================================== #
                         MULTIPLICATION                        #
============================================================== #}
\usage{
.multiplication(x, y)
}
\description{
============================================================== #
                         MULTIPLICATION                        #
============================================================== #
}
