% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{.power}
\alias{.power}
\title{============================================================== #
                             POWER                             #
============================================================== #
Problem:
We don't want to recreate functions
Should we use 'temporal' context?}
\usage{
.power(x, y)
}
\description{
============================================================== #
                             POWER                             #
============================================================== #
Problem:
We don't want to recreate functions
Should we use 'temporal' context?
}
