% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{.remainder}
\alias{.remainder}
\title{============================================================== #
                         REMAINDER                             #
============================================================== #}
\usage{
.remainder(x, y)
}
\description{
============================================================== #
                         REMAINDER                             #
============================================================== #
}
