% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{.subtraction}
\alias{.subtraction}
\title{============================================================== #
                         SUBTRACTION                           #
============================================================== #}
\usage{
.subtraction(x, y)
}
\description{
============================================================== #
                         SUBTRACTION                           #
============================================================== #
}
