% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exe_args_tbl.R
\name{exe_args_tbl}
\alias{exe_args_tbl}
\title{Retrieve the cmdstan arguments available for a compiled executable}
\usage{
exe_args_tbl(code_path)
}
\arguments{
\item{code_path}{Character string describing the path to the Stan code.}
}
\value{
A tibble with rows corresponding to available arguments and columns:
\describe{
  \item{level}{Argument level in the argument hierarchy.}
  \item{parent}{Argument parent in the argument hierarchy (i.e. string that would have to appear before the entry if typed at the command-line).}
  \item{name}{Argument name (i.e. string you'd actually type).}
  \item{description}{Description of the role of the argument.}
  \item{type}{Variable type that the value of the argument should be.}
  \item{valid}{Description of the values that would be valid values for the argument.}
  \item{default}{The default value used by cmdstan if this argument is not otherwise specified.}
  \item{children}{List of the argument's children in the argument hierarchy (if any).}
  \item{order}{Order that the argument appears in the original help output.}
}
}
\description{
Retrieve the cmdstan arguments available for a compiled executable
}
\examples{
\dontrun{
exe_args_tbl('my_model.stan')
}
}
