% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_in_pointer_list}
\alias{is_in_pointer_list}
\title{Check if externaptr object occurs in the given list of externalptrs}
\usage{
is_in_pointer_list(ptr, ptr.list)
}
\arguments{
\item{ptr}{externalptr object}

\item{ptr.list}{list of externalptr objects}
}
\description{
Check if externaptr object occurs in the given list of externalptrs
}
\examples{
library(aria)
ctx <- get_context()
register_ops(ctx, cars)
register_ops(ctx, data.frame)
x <- cpu_tensor(5, dims = 1)
y <- x ** 3
all_ops <- get_all_ops_ptr(ctx)
is_in_pointer_list(aria$xptr,  all_ops)
}
