% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalia.R
\name{marginalia}
\alias{marginalia}
\title{Collect marginalia from aria}
\usage{
marginalia()
}
\value{
a named list of tibbles : `time`, `stdout`, `stderr`
}
\description{
Collect marginalia from aria
}
\examples{
\dontrun{
aria::marginalia()
}
}
