% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{monitor}
\alias{monitor}
\title{Monitor sampling and collect results}
\usage{
monitor(var = "aria_out", as_job = TRUE)
}
\arguments{
\item{var}{Character string describing the desired name of the variable to save to the global workspace.}

\item{as_job}{Boolean value, when TRUE (default) launches RStudio Jobs to convey progress information.}
}
\value{
NULL (invisibly); Side effects: if .
}
\description{
Monitor sampling and collect results
}
\examples{
\dontrun{
#example for setting seed argument
aria::start_sampling(my_data,'my_model.stan')
aria::monitor()
}
}
