% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot}
\alias{plot}
\alias{plot.aria_adjacency_matrix}
\title{Plot soprano adjacency matrix}
\usage{
\method{plot}{aria_adjacency_matrix}(adj_mat)
}
\arguments{
\item{adj_mat}{An object of class `aria_adjacency_matrix`}
}
\description{
Plot soprano adjacency matrix
}
