% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsMaker.R
\name{rdsMaker}
\alias{rdsMaker}
\title{RDS maker}
\usage{
rdsMaker(folder, base_year = NULL)
}
\arguments{
\item{folder}{IMPACT run output folder}

\item{base_year}{Base year for which the Reporting should be run}
}
\description{
RDS maker
}
\examples{
\dontrun{rdsMaker(folder = <model_folder>/OutputFiles/Scenarios)}
}
