% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx-ops.R
\name{reassign_ops}
\alias{reassign_ops}
\title{Workaround for the copy-on-modify case
TODO: Does only tensors are impacted by this problem}
\usage{
reassign_ops(ops)
}
\description{
Workaround for the copy-on-modify case
TODO: Does only tensors are impacted by this problem
}
