% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx_context.R
\name{reset_context}
\alias{reset_context}
\title{Reset context by removing all the objects from that}
\usage{
reset_context(ctx = get_context())
}
\description{
Reset context by removing all the objects from that
}
