% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{Title}
\usage{
sample(
  data,
  code_path,
  num_warmup = 1000,
  num_samples = 1000,
  num_chains = NULL,
  start_seed = 1
)
}
\arguments{
\item{data}{list() (or \code{tibble::lst()}) object containing the data.}

\item{code_path}{Character string describing the path to the Stan code.}
}
\value{
NULL (invisibly); Side effects: background processes are
}
\description{
Title
}
\examples{
\dontrun{
sample(my_data,'my_model.stan')
}
}
