% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{scalar}
\alias{scalar}
\title{A shortcut to create scalar as `aria_tensor`}
\usage{
scalar(x)
}
\arguments{
\item{x}{A single numeric}
}
\value{
An instance of `aria_tensor`
}
\description{
You don't have to use scalar explicitly.
}
\examples{
x <- scalar(2)
x
}
