% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exe_args.R
\name{show_exe_args}
\alias{show_exe_args}
\title{Show the available arguments for a compiled cmdstan exe}
\usage{
show_exe_args(code_path)
}
\arguments{
\item{code_path}{Character string describing the path to the Stan code.}
}
\value{
NULL (invisibly); Side effects: Prints the help message listing all available arguments accepted by the cmdstan exe
}
\description{
Show the available arguments for a compiled cmdstan exe
}
\examples{
\dontrun{
show_exe_args('my_model.stan')
}
}
