% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tensor}
\alias{tensor}
\title{Create `aria` tensor of arbitrary shapes}
\usage{
tensor(data, dims, requires_grad = FALSE)
}
\arguments{
\item{data}{numeric vector, array or data.frame}

\item{dims}{Dimensions}

\item{requires_grad}{is gradient required; logical}
}
\description{
Tensor abstraction, an interface for CPU and CUDA tensors.
}
\examples{
library(aria)
x <- tensor(1:30, c(3, 10))
}
