% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{vec_arith.aria_tensor}
\alias{vec_arith.aria_tensor}
\title{Arithmetic operations
 "+", "-", "*", "^", "%%", "%/%", "/"
Dictionary:
x, y - inputs
result - cached operation result
grad - gradinet from previous result}
\usage{
vec_arith.aria_tensor(op, x, y, ...)
}
\description{
Arithmetic operations
 "+", "-", "*", "^", "%%", "%/%", "/"
Dictionary:
x, y - inputs
result - cached operation result
grad - gradinet from previous result
}
