% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{with_scope}
\alias{with_scope}
\title{Temporarily modifies the used scope name}
\usage{
with_scope(ctx, code)
}
\arguments{
\item{ctx}{An `aria_context` object}

\item{code}{`aria` expressions to execute
with the chosen context}
}
\value{
The results of the evaluation of
the `aria` expressions in the selected context
}
\description{
Temporarily modifies the used scope name
}
