% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_compile.R
\name{check_and_compile}
\alias{check_and_compile}
\title{Check syntax & compile}
\usage{
check_and_compile(code_path)
}
\arguments{
\item{code_path}{Character string describing the path to the Stan code.}
}
\value{
NULL (invisibly)
}
\description{
This
}
\examples{
\dontrun{
check_and_compile(
	code_path = 'stan/my_mod.stan'
)
}
}
\seealso{
Other Model checking & compilation functions: 
\code{\link{disable_rstudio_syntax_compile}()},
\code{\link{enable_rstudio_syntax_compile}()}
}
\concept{Model checking & compilation functions}
