% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Delete temporary files associated with a specific model}
\usage{
clean(code_path, keep_exe = F)
}
\arguments{
\item{code_path}{Character string describing the path to the Stan code.}

\item{keep_exe}{Logical value (default:FALSE) indicating whether to keep the exe folder.}
}
\value{
Returns NULL invisibly
}
\description{
Delete temporary files associated with a specific model
}
\examples{
\dontrun{
clean('my_model.stan')
}
}
\seealso{
Other Cleaning functions: 
\code{\link{clean_all}()}
}
\concept{Cleaning functions}
