% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{coda}
\alias{coda}
\title{Collect output from aria}
\usage{
coda(out_path)
}
\arguments{
\item{out_path}{Character string describing the path to a file where the final sampling output was saved.}
}
\value{
a tibble with samples, plus attributes: `stdout`, `stderr`, `adapt_info`, & `sampling_info`
}
\description{
Collect output from aria
}
\examples{
\dontrun{
p = aria::coda('sampled/sampled.qs')
}
}
