% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{coda}
\alias{coda}
\title{Collect output from aria}
\usage{
coda(out_path)
}
\arguments{
\item{out_path}{Character string describing the path to a file where the final sampling output was saved.}
}
\value{
An instance of the "composition" class, with methods `nc_info()` and `draws()`.
}
\description{
Collect output from aria
}
\examples{
\dontrun{
post = aria::coda('sampled/sampled.qs')
post$nc_info()
post$draws(variables='means') #draws_rvars format
post$draws(groups='generated_quantities') #can specify groups instead
}
}
