% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Compile Stan code and check for runtime errors}
\usage{
compile(code_path, as_job = TRUE)
}
\arguments{
\item{code_path}{Character string describing the path to the Stan code.}
}
\value{
NULL (invisibly); Side effects: an eponymous executable binary is placed in a folder called `aria` (along with some other helper files).
}
\description{
This function first checks if th Stan code has changed in functional ways (i.e. not comments nor whitespace), compiles when such changes exist then runs a simple test for runtime errors.
}
\examples{
\dontrun{
compile('my_model.stan')
}
}
\seealso{
Other Model checking & compilation functions: 
\code{\link{check_syntax_and_maybe_compile}()},
\code{\link{check_syntax}()},
\code{\link{disable_rstudio_syntax_compile}()},
\code{\link{enable_rstudio_syntax_compile}()}
}
\concept{Model checking & compilation functions}
