% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{monitor}
\alias{monitor}
\title{Monitor sampling and collect results}
\usage{
monitor(out_path = "aria/sampled.qs", as_job = TRUE)
}
\arguments{
\item{out_path}{Character string describing the path to which aria will save a ArViz-compliant NetCDF4 file.}

\item{as_job}{Boolean value, when TRUE (default) launches RStudio Jobs to convey progress information.}
}
\value{
NULL (invisibly); Side effects: Launches a background process (as an RStudio job if run from RStudio and \code{as_job} is TRUE) to monitor the output of sampling
}
\description{
Monitor sampling and collect results
}
\examples{
\dontrun{
aria::start_sampling(my_data,'my_model.stan')
aria::monitor()
aria::collect()
}
}
