% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx_context.R
\name{register_ops}
\alias{register_ops}
\title{Create an operation inside the context.
It may be a tensor or a function}
\usage{
register_ops(ctx, r.ops, paired.ops = NULL)
}
\description{
Create an operation inside the context.
It may be a tensor or a function
}
